% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases_handling_functions.R
\name{convert_ensembl_to_reactome_pw_tbl}
\alias{convert_ensembl_to_reactome_pw_tbl}
\title{Map ensembl gene id clone tibble to reactome pathway 
clone tibble.}
\usage{
convert_ensembl_to_reactome_pw_tbl(mutated_gene_tbl, ensg_reactome_path_map)
}
\arguments{
\item{mutated_gene_tbl}{The tibble containing the information
of which ensembl gene id is altered in which patient
 and clone. Can be created with e.g. 
 \code{\link{create_tbl_ent_clones}}.}

\item{ensg_reactome_path_map}{A tibble with all ensembl id's 
and their reactome pathways. Can be loaded with 
\code{data("ensg_reactome_path_map")}.}
}
\value{
The tibble containing the information of which 
pathway is altered in which clone.
}
\description{
For a tibble that contains the information which ensembl 
gene id is mutated in which clone, map the ensembl
gene id to the reactome pathways that contain this gene.
}
\details{
Such a tibble can be generated with e.g. the function 
\code{\link{create_tbl_tree_collection}}. If the altered entities
in the lists were the ensembl gene id's, this function 
can convert the tibble into a tibble with the altered
reactome pathways. It has the columns 'file_name', 
'patient_id', 'altered_entity', 'clone1', 'clone2', ... up 
to the maximal number of clones (Default: until 'clone7'). 
If the mutated entities are ensembl gene id's, they can 
be mapped with this function to the pathways from 
'reactome'. The pathways are from the lowest level of hierarchy.
}
\examples{
data("ensg_reactome_path_map")
mutated_gene_tbl <- 
  dplyr::tibble(file_name=c("pat1.csv", "pat1.csv"),
patient_id=c("1","1"),
altered_entity=c("ENSG00000134086", 
"ENSG00000141510"),
clone1=c(1,0),
clone2=c(0,1))
convert_ensembl_to_reactome_pw_tbl(mutated_gene_tbl, 
    ensg_reactome_path_map)
}
\author{
Ariane L. Moore
}
