% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases_handling_functions.R
\name{ensembl_to_hgnc}
\alias{ensembl_to_hgnc}
\title{Get the hgnc gene symbol for an ensembl gene id.}
\usage{
ensembl_to_hgnc(this_ensembl, all_genes_tbl)
}
\arguments{
\item{this_ensembl}{The ensembl id of a gene.}

\item{all_genes_tbl}{A tibble with all genes ensembl id's and hgnc symbols.}
}
\value{
The matching hgnc gene symbol.
}
\description{
Map a given ensembl gene id to the hgnc gene symbol.
}
\details{
For an ensembl id and a tibble with all genes as input, this function returns the matching
hgnc gene symbol. The tibble with all genes can be generated with
\code{\link{create_ensembl_gene_tbl_hg}}.
}
\examples{
\dontrun{
all_genes_tbl <- create_ensembl_gene_tbl_hg()
ensembl_to_hgnc("ENSG00000134086", all_genes_tbl)
ensembl_to_hgnc("ENSG00000141510", all_genes_tbl)
}
}
\author{
Ariane L. Moore
}
