% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases_handling_functions.R
\name{ensembl_to_reactome}
\alias{ensembl_to_reactome}
\title{Get the reactome pathways for an ensembl gene id.}
\usage{
ensembl_to_reactome(this_ensembl, ensg_reactome_path_map)
}
\arguments{
\item{this_ensembl}{The ensembl id of a gene.}

\item{ensg_reactome_path_map}{A tibble with all ensembl 
id's and their reactome pathways. Can be loaded with 
\code{data("ensg_reactome_path_map")}.}
}
\value{
The pathways that contain this gene as a character 
vector.
}
\description{
Map a given ensembl gene id to the reactome pathways that 
contain this gene.
}
\details{
As input, an ensembl gene id is given as well as the tibble 
'ensg_reactome_path_map'. It can be loaded with 
\code{data("ensg_reactome_path_map")}, and contains the 
ensembl gene id to reactome pathway mappings. The reactome
pathways are from the lowest level of the hierarchy. 
This function returns the reactome pathways for the input gene.
}
\examples{
data("ensg_reactome_path_map")
ensg_gene <- "ENSG00000134086"
ensembl_to_reactome(ensg_gene, ensg_reactome_path_map)
}
\author{
Ariane L. Moore
}
