% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ensg_reactome_path_map}
\alias{ensg_reactome_path_map}
\title{Ensembl gene id's and the Reactome pathways.}
\format{A tibble with 46'141 rows and six variables:
\describe{
  \item{ensembl_gene_id}{the Ensembl gene id as a character}
  \item{reactome_pw_id}{the Reactome pathway stable identifier}
  \item{url}{The url leading to the pathway graph}
  \item{reactome_pw_name}{the name of the Reactome pathway}
  \item{evidence_code}{the evidence code}
  \item{species}{the species}
}}
\source{
The tibble was created as follows: 
library(dplyr);
ensg_path_map_raw <- read.csv("Ensembl2Reactome_homo_sapiens.txt", header = F, sep = "\\t", 
    comment.char = "", check.names = F, skip = 0);
stopifnot(dim(ensg_path_map_raw)[1] == 46141);
stopifnot(dim(ensg_path_map_raw)[2] == 6);
colnames(ensg_path_map_raw) <- c("ensembl_gene_id", "reactome_pw_id", "url", "reactome_pw_name", "evidence_code", "species");
ensg_path_map_raw <- dplyr::as.tbl(ensg_path_map_raw);
ensg_reactome_path_map <- filter(filter(ensg_path_map_raw, grepl("ENSG", ensg_path_map_raw$ensembl_gene_id)), species == "Homo sapiens")
}
\usage{
ensg_reactome_path_map
}
\description{
This is a tibble that contains mappings between ensembl gene id's and reactome pathways. The reactome pathways
are from the lowest level in the hierarchy ("Lowest level pathway diagram / Subset of the pathway"), 
and were obtained by download from the Reactome website (\url{https://reactome.org/download-data}; "ENSEMBL to pathways").
The following commands were used:
wget https://reactome.org/download/current/Ensembl2Reactome.txt; 
cat Ensembl2Reactome.txt | grep "Homo sapiens" > Ensembl2Reactome_homo_sapiens.txt
}
\keyword{datasets}
