% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ensmusg_reactome_path_map}
\alias{ensmusg_reactome_path_map}
\title{Ensembl gene id's and the Reactome pathways
 - for mouse!}
\format{A tibble with 28,630 rows and six variables:
\describe{
  \item{ensembl_gene_id}{the Ensembl gene id as a 
  character}
  \item{reactome_pw_id}{the Reactome pathway stable 
  identifier}
  \item{url}{The url leading to the pathway graph}
  \item{reactome_pw_name}{the name of the Reactome 
  pathway}
  \item{evidence_code}{the evidence code}
  \item{species}{the species}
}}
\source{
The tibble was created as follows: 
library(dplyr);
ensmusg_path_map_raw <- 
read.csv("Ensembl2Reactome_mus_musculus_woOmegaSymbol.txt", 
    header = F, sep = "\\t", 
    comment.char = "", check.names = F, skip = 0);
stopifnot(dim(ensmusg_path_map_raw)[1] == 28696);
stopifnot(dim(ensmusg_path_map_raw)[2] == 6);
colnames(ensmusg_path_map_raw) <- c("ensembl_gene_id", 
"reactome_pw_id", "url", "reactome_pw_name", "evidence_code", "species");
ensmusg_path_map_raw <- dplyr::as.tbl(ensmusg_path_map_raw);
ensmusg_reactome_path_map <- filter(filter(ensmusg_path_map_raw, 
grepl("ENSMUSG", ensmusg_path_map_raw$ensembl_gene_id)), 
species == "Mus musculus")
stopifnot(dim(ensmusg_reactome_path_map)[1] == 28630)
}
\usage{
ensmusg_reactome_path_map
}
\description{
This is a tibble that contains mappings between 
mouse ensembl gene id's and reactome pathways. 
The reactome pathways
are from the lowest level in the hierarchy 
("Lowest level pathway diagram / Subset of the pathway"), 
and were obtained by download from the Reactome website 
(\url{https://reactome.org/download-data}; "ENSEMBL to pathways").
The following commands were used:
wget https://reactome.org/download/current/Ensembl2Reactome.txt; 
cat Ensembl2Reactome.txt | grep "Mus musculus" > 
Ensembl2Reactome_mus_musculus.txt
# and then Ensembl2Reactome_mus_musculus_woOmegaSymbol.txt 
was created from this by just
# replacing the greek 'omega'-symbol in pathway 
"R-MMU-9027604" with the word 'omega'.
}
\keyword{datasets}
