% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling_functions.R
\name{extract_num_clones_tbl}
\alias{extract_num_clones_tbl}
\title{Extract number of clones.}
\usage{
extract_num_clones_tbl(clone_tbl)
}
\arguments{
\item{clone_tbl}{The tibble generated with \code{\link{create_tbl_ent_clones}}.}
}
\value{
A named vector with the number of clones in each patient. The name of each element is the respective
patient_id.
}
\description{
Extract number of clones in each patient.
}
\details{
Given a clone tibble as created with \code{\link{create_tbl_ent_clones}} this function extracts the information,
how many clones there are in each patient. The counted clones will be those with at least one non-zero entry, i.e.
at least one gene/pathway assigned to the clone.
}
\examples{
clone_tbl <- dplyr::as.tbl(as.data.frame(cbind(
                       file_name = c(rep("fn1", 2), rep("fn2", 2)),
                       patient_id = c(rep("pat1", 2), rep("pat2", 2)),
                       altered_entity = c("pw1", "pw2", "pw1", "pw3"),
                       clone1 = c(0, 0, 0, 0),
                       clone2 = c(0, 1, 0, 1),
                       clone3 = c(1, 1, 0, 1),
                       clone4 = c(1, 0, 0, 0))))
extract_num_clones_tbl(clone_tbl)
}
\author{
Ariane L. Moore
}
