% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling_functions.R
\name{get_hist_clon_excl}
\alias{get_hist_clon_excl}
\title{Compute all values of how often gene pairs were clonally exclusive across all trees for a patient.}
\usage{
get_hist_clon_excl(clone_tbl)
}
\arguments{
\item{clone_tbl}{A tibble containing the columns 'altered_entity', and then a column for each clone
in the tumor, e.g. 'clone1', 'clone2', 'clone3'. It also contains the column 'tree_id', which specifies
which tree of the collection of tree inferences was used. This tibble can be generated e.g. from the \code{Cloe} output.}
}
\value{
A list with two vectors: The numbers of how often gene pairs were mutated across trees, and the numbers of
how often they were clonally exclusive. The order of these two vectors is matching, i.e. the ith entry in each
vector refers to the same gene pair.
}
\description{
Compute all values of how often gene pairs were clonally exclusive/all trees for a patient.
}
\details{
It computes a histogram of the following two values: Amomg all gene/pathway pairs in a patient, the number of trees in which 
the both entities of a pair are assigned to a clone at all, and the number of trees in which the pair is clonally exclusive.
}
\examples{
clone_tbl <- dplyr::as.tbl( as.data.frame( cbind(
                     altered_entity = c(paste("gene", seq(1,10), sep = "")),
                     clone1 = c(rep(0,10)),
                     clone2 = c(sample(c(0,1), 10, replace = TRUE)),
                     clone3 = c(sample(c(0,1), 10, replace = TRUE)),
                     clone4 = c(sample(c(0,1), 10, replace = TRUE)),
                     tree_id = c(rep(5, 5), rep(10, 5)) ) ))
get_hist_clon_excl(clone_tbl)
}
\author{
Ariane L. Moore
}
