% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases_handling_functions.R
\name{hgnc_to_ensembl}
\alias{hgnc_to_ensembl}
\title{Get the ensembl gene id for a hgnc gene symbol.}
\usage{
hgnc_to_ensembl(this_hgnc, all_genes_tbl)
}
\arguments{
\item{this_hgnc}{The hgnc gene symbol of a gene.}

\item{all_genes_tbl}{A tibble with all genes ensembl id's and hgnc gene symbols.}
}
\value{
The matching ensembl gene id. In case several ensembl gene id's were found, they
are all returned with ";" as a separator.
}
\description{
Map a given hgnc gene symbol to the ensembl gene id.
}
\details{
For a hgnc gene symbol and a tibble with all genes as input, this function returns the matching
ensembl gene id. The tibble with all genes can be generated with \code{\link{create_ensembl_gene_tbl_hg}}.
}
\examples{
all_genes_tbl <- create_ensembl_gene_tbl_hg()
hgnc_to_ensembl("VHL", all_genes_tbl)
hgnc_to_ensembl("PBRM1", all_genes_tbl)
}
\author{
Ariane L. Moore
}
