% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling_functions.R
\name{merge_clones_identical_ents}
\alias{merge_clones_identical_ents}
\title{Merge identical entities in clone tibble from one patient}
\usage{
merge_clones_identical_ents(clone_tbl)
}
\arguments{
\item{clone_tbl}{The clone tibble as generated with \code{\link{create_tbl_ent_clones}} from one patient.}
}
\value{
The same tibble but in case there were several identical genes/pathways in the same patient with
different clone profiles, their profile will be merged together. This can happen if, e.g. two genes with
different clone profiles are in the same pathway. When mapping them to the pathways, there will be
two identical 'altered_entities' with different clone profiles. These profiles would be merged by this
function because the pathway is affected in the union of clones were the two genes were mutated.
}
\description{
Merge clone profile of identical entities in clone tibble from one patient
}
\details{
Given a clone tibble as created with \code{\link{create_tbl_ent_clones}} from one patient 
and where the entities were possibly mapped from genes to pathways, this function checks whether there
were several entities mapped to the same new entity. If so, the clone profile will be merged.
This can be the case, for instance, of two mutated genes are in the same pathway(s).
}
\examples{
clone_tbl <- dplyr::as.tbl(as.data.frame(cbind(
                       file_name = c(rep("fn1", 4)),
                       patient_id = c(rep("pat1", 4)),
                       altered_entity = c("pw1", "pw1", "pw2", "pw3"),
                       clone1 = c(1, 0, 1, 0),
                       clone2 = c(0, 1, 0, 1),
                       clone3 = c(1, 1, 0, 1),
                       clone4 = c(0, 1, 0, 0))))
merge_clones_identical_ents(clone_tbl)
}
\author{
Ariane L. Moore
}
