% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling_functions.R
\name{pairs_in_patients_hist}
\alias{pairs_in_patients_hist}
\title{Pairs in how many patients histogram}
\usage{
pairs_in_patients_hist(clone_tbl)
}
\arguments{
\item{clone_tbl}{The tibble containing the information of 
which gene/pathway is mutated in which
clone from which patient and in which tree from the collection 
of trees. Can be generated with \code{\link{create_tbl_tree_collection}}
for each patient separately and then appended.}
}
\value{
The tibble that summarizes the number of pairs that
occur in how many patients.
}
\description{
Check in how many patients pairs are mutated in
}
\details{
After having created the tibble with all gene-to-clone 
assignments from all patients and the whole collection
of trees, we're interested in how many patients tha 
pairs are mutated in. This function creats a histogram
that shows in how many patients the pairs are mutated in.
}
\examples{
clone_tbl <- dplyr::tibble("file_name" =
    rep(c(rep(c("fn1", "fn2"), each=3)), 2),
    "patient_id"=rep(c(rep(c("pat1", "pat2"), each=3)), 2),
    "altered_entity"=c(rep(c("geneA", "geneB", "geneC"), 4)),
    "clone1"=c(0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0),
    "clone2"=c(1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1),
    "tree_id"=c(rep(5, 6), rep(10, 6)))
pairs_in_patients_hist(clone_tbl)
}
\author{
Ariane L. Moore
}
